refactor(auth): extract MCP OAuth server use case
